Test encoding to different codecs and file formats
This test verifies that we can write to all supported file formats using
every codec where QMediaFormat::isSupported() returns true. The test
reveals that we are unable to record to Matroska with some codecs, and
that there are issues recording with the H264 codec on macOS ARM.

Note that the tests should later be extended to encode video and audio
simulataneously, as this reveals several issues.

The tests have several TODOs regarding recording audio. These should be
investigated later.

Task-number: QTBUG-126276
Task-number: QTBUG-125956
Task-number: QTBUG-130089
Pick-to: 6.8 6.9
Change-Id: Ib062dd7c0c301e631f9f55c7a9069b0d9960f121
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>