Don't require audio format with QMediaFormat::isSupported
QMediaFormat::isSupported now returns true when video format is
supported even if audio format is unspecified. This enables use of
isSupported to check if a codec is supported when encoding or
decoding video without audio.

In this patch we also add tests for QMediaFormat::isSupported for the
FFmpeg media backend with video codecs.

Fixes: QTBUG-130478
Pick-to: 6.8 6.5
Change-Id: I2ecd88a908c5dd7ec635de3b8210850f1c62765d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>