QDarwinAudioSink: handle AudioOutputUnitStart failure
When device is locked, AudioOutputUnitStart can fail with "cannot start"
error. According to the docs, this means that app's category is
inconsistent with background mode or missing entry in the Info.plist.
None of this applies to our case - we are not recording, thus don't
need Info.plist entry and adding background audio mode to
entitlements does not help either. So let's check the application
state change and unit start failures, and re-start it later if
needed.

Pick-to: 6.8
Fixes: QTBUG-114674
Change-Id: I2f4c4fa43f053a756a037966d8c17d4d3714a90f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>