Reset QSoundEffect's audio buffer when changing source
The audio buffer was not reset when changing the audio source, causing
the old audio buffer to be reused, unless the FFmpeg media backend is
present.

This patch fixes this issue by forcing the audio buffer to be recreated
when the source is changed. The minimal QML sound effect example is
updated to make this scenario easier to test.

Amends: f0016c4763361292c3eaf153e184b9b110f44c61

Fixes: QTBUG-130668
Pick-to: 6.5 6.8
Change-Id: Iba10cfb6d79a6411f9d1bc10d3fd8c4925e9fcb1
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>