Only call find_package when configuring test as standalone CMake project
When configuring media-metadata as standalone test or as part of the Qt
Multimedia build using QT_BUILD_MANUAL_TESTS, find_package calls are
unnecessary and can cause configure error due to a design issue in how
CMake handles the static mock plugin.

This patch fixes the issue by not calling find_package if building Qt or
if building standalone tests.

Pick-to: 6.8 6.5
Change-Id: I15f7be278d106c38faa1141d34cde0082b02de75
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>