Add 'minimal' manual QML test for audio recording
This test is a starting point for playing with recording audio using
QML. It sets up a media session with one audio input and a recorder. The
test also allows selecting the input audio device, and can for example
be used to test what happens when changing audio inputs or disonnecting
audio inputs.

Task-number: QTBUG-123073
Pick-to: 6.8
Change-Id: I95e9604c6434ff10e9e18141c0a7dba6bf93eb8d
Reviewed-by: Mikko Hallamaa <mikko.hallamaa@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>