Change device on QComboBox::currentIndexChanged instead of -::activated
The audio output example behaves more intuitively if it always switches
to the default system audio output whenever the default system audio
output changes. While the device combobox updates accordingly, the audio
sink does not. This discrepancy arises because the QAudioSink is only
reinitialized when the audio device is changed manually, not
automatically when the default system audio output changes.

To address this issue, we modify the device change handling to occur on
QComboBox::currentIndexChanged instead of QComboBox::activated. The
latter is only emitted when the user selects an index in the combobox,
whereas QComboBox::currentIndexChanged is emitted whenever the selected
item changes, either by user action or programmatically.

Fixes: QTBUG-119506
Pick-to: 6.8 6.5
Change-Id: I91a9e2823876336f17e220d5548afa89c5573157
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Lars Sutterud <lars.sutterud@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>