Query QRhi for video format texture support
We hard-coded video plane texture formats for each pixel format, which
was problematic when the rhi used didn't support the format we
preferred, causing uninitialized textures and unusable video rendering.

This patch adds internal Format enum for texture descriptions, which can
be converted to QRhiTexture formats when an rhi instance is available.
This allows centralized handling of fallbacks when preferred texture
format is not supported by rhi.

Instead of passing boolean to determine alpha shader usage, we should
just pass a pointer to rhi instance being used in order to get full
picture of the capabilities of current rhi, and the eventual texture
format used for the video textures.

Pick-to: 6.9 6.8
Change-Id: I2d3cd5c141d71c97b889a0abf88ea920395ab508
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>