Move free functions that operate on codecs to their own file
The qffmpeg_p.h is getting a bit hard to read. This patch makes it a bit
slimmer by moving functions that are related to AVCodec to its own
files. We also hide implementation functions that are not needed
elsewhere inside the cpp file.

The getCodecChannelLayouts function is adjusted a bit, to avoid
conditionally compiling in two different versions. Instead we modify
its return type based on the QT_FFMPEG_HAS_AV_CHANNEL_LAYOUT flag.

adjustChannelLayout is modified similarly, to have only one declaration,
and one definition.

Task-number: QTBUG-131270
Pick-to: 6.8 6.5
Change-Id: I6263e3c6db095ea9ea18d985c2a4be0625d72956
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>