Return QSpan instead of raw pointers from Codec class
Several of Codec's functions returned raw pointers, and it was up to the
caller to determine what the sentinel stop value was to be able to know
how many elements the array contained.

This patch mitigates this issue by using QSpan to return views onto
FFmpeg arrays, and the responsibility for determining the number of
elements in the arrays is moved inside the Codec class. Callers no
longer need to know which value is used to terminate the array, which
makes the Codec API safer to use.

Task-number: QTBUG-131270
Pick-to: 6.8
Change-Id: I7ee5d67eb176ecfcc6e370e85ff61157725aac58
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>