Use QT_CLOSE to close handles, and avoid designated initializers
This makes it possible to compile the GStreamer backend with msvc, where
designated initializers require C++ 20. qt_safe_close is not available
with msvc, so we need to switch to QT_CLOSE isntead. GStreamer is not
supported on Windows, but for Qt multimedia development it can be useful
to be able to compile the backend with the Microsoft compiler.

Pick-to: 6.8
Change-Id: I5c016b5f1bb86526a25bc9dea7abcb8d117c6ecd
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>