Replace hwConfig(int) with a new hwConfigs() function to QFFmpeg::Codec
This new function returns a vector of items, and results in cleaner code
on the calling side than the old hwConfig(int) function, and gives a
better abstraction that is easier to iterate over than the old version.

Most FFmpeg codecs do not have any HW configuration, and this function
return an empty vector most of the time. With the help of return value
optimization, the overhead of accumulating items in this vector become
very low.

Task-number: QTBUG-131270
Pick-to: 6.8
Change-Id: Iabb7fa21bdf01edcac7cf6ab1f9163032726875d
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>