Remove unused QGstreamerSyncMessageFilter and related functions
The QGstreamerSyncMessageFilter is not in use. We see this because
QGstBusObserver::installMessageFilter and
QGstBusObserver::removeMessageFilter functions were never called with
a QGstreamerSyncMessageFilter instance.

This patch removes the QGstreamerSyncMessageFilter interface and
functions and member variables that were using it. We also remove calls
to gst_bus_set_sync_handler because we no longer need to respond to sync
callbacks.

Pick-to: 6.8
Change-Id: I3390d636524919cd637f061973896ba090fd6254
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>