FFmpeg, Darwin: Don't crash during texture-conversion
Previously we would pass MTLPixelFormatInvalid into
CVMetalTextureCacheCreateTextureFromImage during AVFrame conversion
when we are unable to map the incoming pixel format to a valid one.
This causes a segfault. This would happen when the incoming
plane had format RED_OR_ALPHA8.

This patch makes us early return if we encounter such a scenario.
Texture conversion will fail but without causing a crash and
rendering will revert to not using zero-copy transfer.

Task-number: QTBUG-131530
Pick-to: 6.8 6.5
Change-Id: I904c4a348a43637bc1baf6c1eb42b9399e8324c1
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>