Docs: Incorrect images are used for QML Media Player example on the web
This commit updates one of the affected image and renames both affected
images. The wrong images on web were displayed because the are another
repositories that have images with the same names. During the
deployment, the images from the Qt Multimedia module were overwritten
with other images with the same names. Adding more context to the image
name makes the names unique and thus resolves the issue.

Fixes: QTBUG-131567
Pick-to: 6.8
Change-Id: Ie3602da9bb18e0e8124b6ec9bddf8c6d288aba1e
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>