QCamera, Android: Add more commenting to Java code
It's currently fairly unclear to a reader what the code does, unless
the reader is already very intimate this code. Many properties are
used in ways that are not immediately obvious and this is hurting
maintainability.

This patch adds comments to several of the control properties
to make it more clear what purpose they serve and how they should be
used.

No functional changes.

Pick-to: 6.8 6.5
Change-Id: I211718d569dd2f1be5b22fdd6e802566ae7073e9
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>