Android, Camera: Improve error handling in still-photo finalizer
Bug report has evidence there is an invalid null access on
'mCaptureSession'. Generally this should not be possible because
mCaptureSession is immediately set to null after calling .close(), which
should flush all pending capture-related callbacks.

This patch improves the error-handling in the relevant code by catching
null-access exceptions. This does not solve the underlying problem, but
stops us from crashing when it occurs.

Fixes: QTBUG-130901
Pick-to: 6.9 6.8 6.5
Change-Id: I32129e31614b7e697afc62b62b48487ddc1c2dcd
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>