Add the endOfStreamPolicy property to VideoOutput
In Qt5 we had FlushMode for a simialar purpose.
The term Flush doesn't work well for this kind of functionality,
so it's better to call it according to what it's doing.

We want to have this API in qml only, because users have
an opportunity to filter frames through their own video sink in c++.

The method clearOutput() allows users to manually clear the last frame
if it has been kept and not needed anymore.

[ChangeLog] Added the qml property VideoOutput.endOfStreamPolicy
allowing users to keep the last frame at the end of the stream.
Added the qml method VideoOutput.clearOutput() allowing users
to manually clear the last kept frame.

Task-number: QTBUG-110310
Change-Id: I238feb1592d5d055b2e04804e13a60b8c5307c78
Reviewed-by: Alexei Cazacov <alexei.cazacov@qt.io>
Reviewed-by: Nils Petter Skålerud <nils.petter.skalerud@qt.io>
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>