Android, FFmpeg: Comment threading considerations in QAndroidCamera
Currently it's a bit unclear what functions are called from which
threads at any given time. This makes it harder to maintain and review
code patches in the codebase.

This patch introduces some comments on what threads are allowed to
run certain specific functions. It also introduces some TODO
comments on possible race conditions that might lead to problematic
bugs in the future.

No functional changes.

Pick-to: 6.8 6.5
Change-Id: I6ecc7c677fc794ff170974a53b7f781cc23d49f3
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>