Log FFmpeg version and license information during startup
This patch adds logging of the FFmpeg version being used and its
license on application startup.

This can help developers and end users to see which version of FFmpeg
the application uses, and which licensing terms the application is
under.

We also change from qDebug() to qLcInfo() for logging supported hardware
configurations, because QT_NO_DEBUG is designed for disabling asserts,
not for controlling debug log output. We use qLcInfo() instead of
qDebug() to show the log by default as it is with qDebug().

Task-number: QTBUG-131759
Pick-to: 6.8
Change-Id: I2440fceaf0180a722235ebfbcdff2fc601bd4c13
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>