Switch to default audio device when audio devices change in mediaplayer
The mediaplayerexample stops playing if the current audio device is
disconnected. This can be fixed by always switching to the default audio
output whenever the media devices change.

If the last audio device is not disconnected, the video playback still
stops, but it is recoverable by pushing play after connecting the audio
device again.

Fixes: QTBUG-119355
Pick-to: 6.8
Change-Id: Ia9bae9af366272217fe2c1d6cc009fb24a4ad2f2
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>