QML Video Example: Refactor into two QML modules
This change fixes an issue on macOS and windows where the example
doesn't work because some types cannot be found. It also refactors
the example with these changes:

- Merge QML module peformancemonitor into frequencymonitor

- Avoid manual calls to qmlRegisterType by relying on automatic type
registration via qt_add_qml_module(). This was done by moving all source
files to their respective QML module folder, and adding
QML_ELEMENT to the C++ class FrequencyMonitor.

- Remove logic to deactivate frequency and performance monitoring
through the cmake flags FREQUENCYMONITOR_SUPPORT and
PERFORMANCEMONITOR_SUPPORT. This functionality, together with its QML
module based implementation, is highlighted in the example
documentation, and being able to deactivate it is not necessary. The
logic relied on qmlRegisterType, making it convenient to remove it now.

- Delete performancemonitor.c/h by moving all cmd line parsing to
main.cpp.

- Update qdoc quotes to reflect changes.

- Merge external file video-qml-paint-rate.qdocinc into qmlvideo.qdoc

Note, this change only fixes the example for cmake, qmake doesn't
work.

Fixes: QTBUG-131286
Change-Id: Ibd183c180439b79277d914e6fe3dd4c50707b857
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>