Darwin audio sink: set MixWithOthers option explicitly
While documentations says that having 'Ambient' category on audio session
implies MixWithOthers - this does not seem to be true: switching silent
mode to 'ON' on a devices silences the audio, but only with FFmpeg
backend, which relies on QDarwinAudioSink. In case of 'darwin' backend
the behavior is different though - audio playback continues (and it's
audible). Documentation in our code also implies we rely on such behavior,
so work the problem around by setting the appropriate category
and options.

Fixes: QTBUG-130340
Change-Id: I1383a4416e864389f87f233deafe1caae8caba5a
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>