AVFMediaPlayer/Observer: reset the 'retain' property in 'dealloc'
While 'retain' property does some memory management for us (releasing
the previous and retaining the new value in an assignment e.g.) it
is not releasing in the end, unless we manually set it to nil.
The leak is quite visible just from the fast growing app's memory
footprint but even better - the 'leaks' tool immediately points
on many leaked AVPlayerItems.

This affects only 'darwin' backend, not 'ffmpeg'.

Task-number: QTBUG-129692
Pick-to: 6.8
Change-Id: Idcceb27f413f2fdc3f7c3d90da7786a65abf5437
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>