Android: Add a limit for screen capture and camera frames
We need to handle the limit resulting from [0]ImageReader.
ImageReader does not allow to take more frames than the maxImages parameter
used during creation. We need to control the number of unclosed images
to avoid IllegalStateException.

This mechanism is used by QAndroidCamera and QAndroidScreenCapture.

The idea is to control the number of frames acquired. Once the limit is
reached, we will start copying the given frames. This will allow closing
images and taking more without the mentioned exception.

For this purpose, the QAndroidVideoFrameFactory class has been added. It
will control the number of frames, decide on copying/reusing memory and
create QVideoFrames.

[0]https://developer.android.com/reference/android/media/ImageReader

Task-number: QTBUG-112855
Pick-to: 6.9 6.8
Change-Id: I18c39a08bc46aad14ed434f2dafd92f3603995d3
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>