AVFVideoBuffer: do not try to create a texture with an invalid format
AVFVideoBuffer converts only a limited number of formats from QRhi to
Metal formats. An attempt to use 'MTLPixelFormatInvalid' for such
'unknown' case or unsupported format ends up in Metal API validation
failure which also makes an app to abort abnormally. If we know
the pixel format is not valid, no need to try to create a texture.
Also, handle RED_OR_ALPHA8.

Fixes: QTBUG-131882
Pick-to: 6.8
Change-Id: I2bcc7f2f740c778de22a991d6db5e08df9580a43
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>