Select conversion type based on scale direction
This commit introduces a dynamic selection of the scaling algorithm
based on the scaling direction to address a crash on Android emulator
caused by the FFmpeg function ff_hcscale_fast_c.

The crash occurs when using the SWS_FAST_BILINEAR scaling algorithm
during anisotropic or upscaling scenarios. To solve this issue
SWS_BICUBIC algorithm is selected when at least one dimension is
upscaled.

Fixes: QTBUG-130898
Pick-to: 6.8 6.5
Change-Id: Iee92f51217da343ac166d5a7e86028bdfb0bf20f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>