QAudioDevice: Make comparison only rely on id and mode
Previously comparison would take into account all the properties of
QAudioDevice, including isDefault, the d pointer etc. These are values
that are cached, and the values may not always reflect the reality
of the current device state.

This patch makes it so that comparison of QAudioDevice only takes into
account the id and output/input nature of the device, making it possible
for a more generalized comparison.

[ChangeLog][Multimedia][QAudioDevice] Comparison of QAudioDevice now
only relies on .id() and .mode() properties

Task-number: QTBUG-131927
Pick-to: 6.9 6.8 6.5
Change-Id: Ieb4aa51688493f19dc06da8464ddf780e0e7986c
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>