Keep textures in rendering slots instead of frames
In many cases, the textures refer to frames, we used to keep
textures anyway.
The patch will allow not to maintain mapped rhi textures
in FFmpeg frames. Video frames can be removed in any thread,
so we should transfer (copy) hw data from frames to textures
as much as we can.

The patch is not a finished work, it's a step to improve the
rendering pipeline.

Task-number: QTBUG-131918
Pick-to: 6.8 6.9
Change-Id: I2162f3ce280b9d666fc16f39e68ce5d461a0f00e
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>