Tests: Update QScreenCaptureBackend test to work on Android
When QtScreenCaptureService is stopped and mService is null, there are
two possibilities:
- the service has not been started at all;
- the service has not been bound yet.

While the first case is fine, the second one may be dangerous (since the
service will not be cleaned up properly). To avoid the second case,
synchronization has been added.

There are also few changes that were added to test for Android:
  1. ScreenCapture feature needs declaration of foregroundServiceType in
     AndroidManifest.xml file. The file was added;
  2. Android is not a Window System. When calling setGeometry() on the
     main widget, it will be displayed at the beginning of the screen.
     That is why geometry is remembered and set later in the paintEvent;
  3. Android is trying to be smart with producing new frames. It is
     really lazy with grabbing frames - when nothing changed on the
     screen. To force the Screen Grabber to work the 10x10 blinking
     rectangle was added
  4. Calculate height of status bar. As the whole screen is recoded, the
     height of status bar has an impact on pixel position calculations

Tests works locally on emulator, but there is still the security popup
that need to be manually accepted. There is adb command that allows to
accept it automatically, but currently there is no way to call it from
androidtestrunner.

There are also still some issues with the "Pixel perfect" color control.
On some devices (with pixelRatio, e.g. 2.625) there are still incorrect
pixel position calculations.

Fixes: QTBUG-112855
Pick-to: 6.9 6.8
Change-Id: I07269bf005fc0fbc3790c1e00c9aa881e3ff25bf
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>