Improve hierarchy in video frame textures
Textures from memory rely on the previous texture from the texture pool,
and it does dynamic cast to check the texture is of the proper type.

The CR makes the logic more robust: it decouples the texture
classes by their purposes in order not to mix one with another upon
casting.

Task-number: QTBUG-131918
Pick-to: 6.8 6.9
Change-Id: I422db9369c63d2aea364919f5daaf72fe2e85578
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>