GLES2: Use RGBA8 as fallback when RG8 is not supported by rhi backend
This patch adds RGBA8 as a fallback texture format for two-component
planes (in NV12 and NV21 pixel formats). This packs two pairs of UV
values together. NV12 and NV21 textures that use alpha component to
Y-value are updated to handle unpacking and blending UV values.

This also requires the addition of rhi-aware plane scaling in texture
descriptions, as packing causes textures to be half as wide, so the
scaling factor should be doubled to compensate if RGBA8 texture is used
for two-component planes. This patch adds methods to query rhi
capabilities to determine horizontal plane scaling, as well as a helper
method to get plane size directly from texture description.

Fixes: QTBUG-131114
Pick-to: 6.9 6.8
Change-Id: Iffcef4df8d14e4dd100a6103b7c9955f1451fa2f
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>