Darwin, FFmpeg: Don't expose camera formats that are not usable
Currently we expose camera videoformats that don't work with the active
media backend. An example is the YUYV format, for which we are
unable to translate into a corresponding FFmpeg pixel format. Trying
to activate such a format leads to warnings and graphical artifacts.

This patch introduces a helper function parameter that can tell us
ahead of time whether a format will work for a capture session. It
uses this function to not expose such formats for QCameraDevice.

Fixes: QTBUG-126305
Pick-to: 6.9 6.8 6.5
Change-Id: Ic81a717ea9a4d4cf5beb039b516fc74795cc7d5b
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>