Add manual test: qml-multiple-players
This new manual test is a QML app consisting of a list of
multiple MediaPlayers. Each player is paired with a VideoOutput
and an AudioOutput. Players can be added or removed with dedicated
buttons.

There are also individual and global buttons to show/hide
the VideoOutputs, mute the AudioOutputs, control playback, skip to
the next media source and toggle looping on/off.

The media sources are defined in a list of strings called
"sources" in the top of qml-multiple-players.qml. Two small
test files are included in the source code.

Change-Id: Ia67c6f552be45d47634945da44e6a13247d3583c
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>