QMediaFormat: provide consistent ordering of codecs/formats
Supported codecs/formats are returned in arbitrary order. This smells a
bit, since different invocations return different orders.
It is not uncommon that this list is directly used to populate gui
elements (like e.g. the qt examples are doing).
We can easily ensure a well-defined order by collecting formats/codecs
into a std::set instead of a QSet.

Pick-to: 6.8 6.9
Change-Id: I7e3bf1409864cf7c057e5f719e16761b04fca7e1
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>