Don't assert in AudioRenderer::pushFrameToBufferOutput with empty frame
The first frame pushed to AudioRenderer::pushFrameToBufferOutput may be
an empty frame. In this case, the resampler has not been created yet.

This patch prevents asserts in such cases by only asserting when a valid
frame is provided. This way, we don't assert if the custom audio input
source only supplies one audio buffer which is empty.

Pick-to: 6.9 6.8 6.5
Change-Id: I9485046f7a5633f12337f4e1d74eff999033f4ef
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>