Android, FrameFactory: Apply rotation to frame surfaceFormat
Currently the frame-factory will apply rotation directly to the
QVideoFrame object, rather than to the internal QVideoSurfaceFormat.
This is problematic when forwarding this frame to
qImageFromVideoFrame(), which only considers the surfaceFormats
rotation. This causes a bug during camera image captures on Android,
where no rotation is applied to the resulting QImage during conversion.

This patch modifies the QAndroidVideoFrameFactory to apply the rotation
directly to the surfaceFormat rather than the QVideoFrame object.

Fixes: QTBUG-132779
Pick-to: 6.9 6.8 6.5
Change-Id: I5bf595363613c4f5f3173cbedaf634a06bee5c57
Reviewed-by: Pavel Dubsky <pavel.dubsky@qt.io>
Reviewed-by: Bartlomiej Moskal <bartlomiej.moskal@qt.io>