Android: Do not use native MenuBar on Android in Camera example
Android MenuBar does not work with Qt on Android 10 or higher.

The problem with native MenuBar is directly related with setting
Theme_DeviceDefault_DayNight. It looks like the Theme does not include
an ActionBar.

The solution in this case could be to use Theme.AppCompat.DayNight
(but requires AndroidX). The current solution is a workaround until
Android X support is available in Qt.

Fixes: QTBUG-112952
Pick-to: 6.9 6.8
Change-Id: I86df18748c23b26ba148efc85950a30bbeb85dd4
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>