Read texture size from texture, don't rely on possibly out of sync size
When mapping a QD3D11TextureVideoBuffer used for Windows screen capture,
the mapped size was determined using a size variable that could be out
of sync with the actual texture size.

This patch fixes this by obtaining the mapped size from the texture that
is being mapped. This can potentially prevent out-of-bounds memory
access, although this has never been observed.

Task-number: QTBUG-129708
Pick-to: 6.5 6.8 6.9
Change-Id: I437148eae3ea4abc4a575c5c15deb11e0e3d1713
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>