Refactor DXGI screen capture to calculate screen size using DXGI
Windows API function GetMonitorInfoW was previously used to calculate
the size of the screen. This function returns the size of the screen
taking rotation into account. Since the DXGI desktop duplication APIs
does not change the size of the grabbed texture even if monitor is
rotated, a rotated monitor would give a size mismatch, causing wrong
aspect ratio with rotated monitors.

Task-number: QTBUG-129708
Pick-to: 6.5 6.8 6.9
Change-Id: Ieb0b9eb41e2435d9443cc4d357b15d89984400c4
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>