Multimedia, iOS: Stop build if no FFmpeg frameworks are found
Currently the build will silently fail during configuration when
deploying iOS FFmpeg frameworks if the framework files are not found.
This causes the FFmpeg to fail linkage when building Qt applications.
This is particularly problematic if not caught during official
release builds.

This patch makes the CMake configuration fail if we explicitly request
FFmpeg deploument but can't find the frameworks. This should help us
detect when release builds are not deploying FFmpeg correctly.

Pick-to: 6.9 6.8
Task-number: QTBUG-132778
Change-Id: I5a9db98aadb27183ba7724eb85b2c618d1502c22
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>
Reviewed-by: Jøger Hansegård <joger.hansegard@qt.io>