Improve logging and fix leak in in QGstreamerIntegration ctor
When reviewing CI logs and bug reports it is useful to see which
GStreamer version is being used. This patch changes the qCDebug output
to qCInfo, to make it log the version by default. The message is also
harmonized with the string logged by the FFmpeg media backend.

This patch also fixes a tiny memory leak (we leaked the version string),
and removing an aliasing variable.

Pick-to: 6.9 6.8 6.5
Change-Id: I84939f21d4fb1a30928aaeea5b43bb2c43656ee8
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>