AudioRecorder: Filter sample rates using a closed interval
Filtering sample rates using an open interval filters out all sample
rates if minimum and maximum sample rates are the same. This prevents
the AudioRecorder from working on Windows, where the audio device
reports that both minimum and maximum sample rate is 48000 Hz.

This patch fixes this by allowing sample rates that are equal to the
min and max sample rates.

Task-number: QTBUG-132532
Pick-to: 6.8 6.9
Change-Id: Ia2309faded74885e260ae586f3e719fd072fd6b2
Reviewed-by: Tim Blechmann <tim.blechmann@qt.io>