Always use stereo channel layout with FFmpeg mp3_mf codec
The FFmpeg mp3_mf codec does not report any restrictions on channel
layouts, but fails when attempting to encode more than two channels. It
also fails to encode with front center (mono) layout.

We work around this by providing the missing channel layout restriction.
This fixes a problem where recording audio fails to record to MP3 with
4 channel microphone arrays.

The Codec::channelLayouts() function is currently only used with
encoding, not with playback.

Fixes: QTBUG-124131
Pick-to: 6.9 6.8
Change-Id: I3653b32045266885441ea644054a1e5e6ab573d7
Reviewed-by: Artem Dyomin <artem.dyomin@qt.io>